refactor: pages 폴더와 components 폴더로 각각의 관심사별로 분리
관심사별로 파일 분리
그에 따른 경로 변경
