+++ src/components/Categories/CategoriesContainer.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from './slice';
+} from '../../slice/slice';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
