+++ src/components/Regions/RegionsContainer.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './slice';
+} from '../../slice/slice';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
