+++ src/pages/LoginPage/LoginFormContainer.jsx
@@ -7,9 +7,9 @@ import {
   changeLoginField,
   requestLogin,
   logout,
-} from './slice';
+} from '../../slice/slice';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
