+++ src/pages/RestaurantPage/RestaurantContainer.jsx
@@ -3,16 +3,16 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import RestaurantDetail from './RestaurantDetail';
-import ReviewForm from './ReviewForm';
-import Reviews from './Reviews';
+import ReviewForm from '../../components/ReviewForm/ReviewForm';
+import Reviews from '../../components/Reviews/Reviews';
 
 import {
   loadRestaurant,
   changeReviewField,
   sendReview,
-} from './slice';
+} from '../../slice/slice';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function RestaurantContainer({ restaurantId }) {
   const dispatch = useDispatch();
