+++ src/pages/RestaurantPage/RestaurantContainer.test.jsx
@@ -82,7 +82,7 @@ describe('RestaurantContainer', () => {
           fireEvent.change(getByLabelText(label), { target: { value } });
 
           expect(dispatch).toBeCalledWith({
-            type: 'changeReviewField',
+            type: 'application/changeReviewField',
             payload: { name, value },
           });
         });
