+++ src/pages/RestaurantsPage/RestaurantsPage.jsx
@@ -4,13 +4,13 @@ import { useHistory } from 'react-router-dom';
 
 import { useDispatch } from 'react-redux';
 
-import RegionsContainer from './RegionsContainer';
-import CategoriesContainer from './CategoriesContainer';
+import RegionsContainer from '../../components/Regions/RegionsContainer';
+import CategoriesContainer from '../../components/Categories/CategoriesContainer';
 import RestaurantsContainer from './RestaurantsContainer';
 
 import {
   loadInitialData,
-} from './slice';
+} from '../../slice/slice';
 
 export default function RestaurantsPage() {
   const history = useHistory();
